requirements.txt
fastapi
uvicorn[standard]
firebase-admin
requests
langchain-openai
openai
langchain-community
faiss-cpu